Refactored how the TUD logo is included in the custom.css:
Instead of requiring the SVG file to be in the same directory as custom.css, the SVG is injected as a byte string into custom.css by means of a common.css, which is automatically generated in start_notebook().

Also added a test script test_themed_notebooks.py in tests.